Refactor: move simple interpolation regex to `grammar_parser.py`
This ensures that grammar-related code and tests are centralized.
